requirements.txt
pandas
requests
python-dotenv
rich
fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
bcrypt>=4.0.0
python-multipart
sqlalchemy>=2.0
beautifulsoup4
lxml
playwright
httpx